started on history page

+++ app/src/main/java/com/example/eurdpayment/Algorand/Models/Asset.kt
@@ -1,3 +1,3 @@
 package com.example.eurdpayment.Algorand.Models
 
-data class Asset(val name: String, val id: Int, val decimals: Int)
+data class Asset(val name: String, val id: Long, val decimals: Int)
